LUN-362: Fix ts check
