+++ src/services/__tests__/AsyncStorage.spec.ts
@@ -158,39 +158,4 @@ describe('AsyncStorage', () => {
       expect(updatedUserVocabulary).toHaveLength(0)
     })
   })
-
-  describe('calculateScore', () => {
-    const getDocumentsWithResults = (
-      numberOfTries: [number, number, number, number],
-      results: [SimpleResult, SimpleResult, SimpleResult, SimpleResult]
-    ): DocumentResult[] => {
-      const documents = new DocumentBuilder(4).build()
-      return documents.map((document, index) => ({
-        document,
-        result: results[index],
-        numberOfTries: numberOfTries[index],
-      }))
-    }
-
-    it('should calculate score correctly for different number of tries', () => {
-      const score = AsyncStorage.calculateScore(
-        getDocumentsWithResults([1, 2, 3, 3], ['correct', 'correct', 'correct', 'incorrect'])
-      )
-      expect(score).toBe(4)
-    })
-
-    it('should calculate score correctly for best result', () => {
-      const score = AsyncStorage.calculateScore(
-        getDocumentsWithResults([1, 1, 1, 1], ['correct', 'correct', 'correct', 'correct'])
-      )
-      expect(score).toBe(10)
-    })
-
-    it('should calculate score correctly for bad result with similar results', () => {
-      const score = AsyncStorage.calculateScore(
-        getDocumentsWithResults([3, 3, 3, 3], ['similar', 'incorrect', 'incorrect', 'incorrect'])
-      )
-      expect(score).toBe(0)
-    })
-  })
 })
